Improved search and related channels logic
#Changes
-----------------------------------
##Index.html
Added links to the push out hamburger menu, still thinking  about what
exactly to put in here.
-----------------------------------
##Instant.js
Added ```updateRecommendedStreams()``` function
This function was initially made when there is only one channel pulled
from the initial API call in ```searchStreams();```. But the query is
different from the first function call. In this case I would remove all
numbers from the query with a regex "/[^a-zA-Z]+/g". Then I only update
the recommended streams/channels with the new query. This can hopefully
fix issues with queries being too specific to have related channels not
appear.

Another case when this function is called is when someone clicks on a
recommended channel and I need to update only the recommended channels
and nothing else.

```hasRelatedChannels()``` now takes a query parameter as well as the
data object. This allows me to update the recommended channels by
passing the updated query

```recommendedUpdate()``` now uses ```updateRecommendedStreams()```
instead of searching for another channel, this would break the clicking
recommended channels in the past.