Added repo svg
These are probably going to be the last two commits I make because of the twitch instant search update a while ago. I feel bad leaving the project in this incomplete state, but I have lost a lot of motivation to work on it. I did learn a lot making it though, and it was a good experience.